refactor: tidy up profiles.ts by extracting watcher registration
Move updateInheritedSettingsOnProfileChange, registerCurrentProfileSaveWatcher,
registerParentProfileSaveWatcher, and their shared createFileWatcher helper out
of profiles.ts and into a new src/profileWatchers.ts module.

profiles.ts now only contains read/write/merge logic for profile settings and
extensions, while profileWatchers.ts owns all file-system watcher wiring that
reacts to profile switches and settings.json edits. The selfWriteTracker
instance in profiles.ts is now encapsulated behind an exported
isManagedFileSelfWrite(filePath, content) helper instead of being exposed
directly, giving profileWatchers.ts a narrower interface to depend on.

This is a pure refactor with no behavioural change: extension.ts and the
integration test suite only needed their imports split across the two
modules. All 32 unit tests and 10 integration tests continue to pass
unmodified.

Checks off the last "Future Plans" item in the README: tidy up profiles.ts.
