+++ src/profiles.ts
@@ -2,7 +2,6 @@ import * as vscode from "vscode";
 import * as path from "path";
 import * as fs from "fs/promises";
 import { parse } from "jsonc-parser";
-import { createDebouncedTrigger } from "./debounce";
 import {
   buildInheritedSettingsBlock,
   findTabValue,
@@ -14,7 +13,6 @@ import {
   mergeInheritedExtensions,
   removeInsertionBoundarySetting,
   removeTrailingComma,
-  resolveParentSettingsPaths,
   sortSettings,
   splitRawSettingsByClosingBrace,
   stripInheritedExtensions,
@@ -33,8 +31,8 @@ const selfWriteTracker = new SelfWriteTracker();
 /**
  * Writes `content` to `filePath` and records it with {@link selfWriteTracker}
  * so that file watchers set up to react to external edits of this file (see
- * {@link registerCurrentProfileSaveWatcher}) can recognise and ignore the
- * change this write is about to cause.
+ * `isManagedFileSelfWrite`) can recognise and ignore the change this write is
+ * about to cause.
  */
 async function writeManagedFile(
   filePath: string,
@@ -45,23 +43,18 @@ async function writeManagedFile(
 }
 
 /**
- * Creates a {@link vscode.FileSystemWatcher} that watches a single file at an
- * absolute path.
- *
- * `vscode.workspace.createFileSystemWatcher` does not reliably report
- * changes for a plain absolute path string when that path lies outside of
- * every open workspace folder — which is always true for profile files,
- * since they live under VS Code's global storage/user directory rather than
- * a workspace. Using an explicit {@link vscode.RelativePattern} with the
- * file's parent directory as its base avoids this problem.
- * @param filePath Absolute path to the file to watch.
+ * Reports whether the most recent write to `filePath` (via
+ * {@link writeManagedFile}) wrote exactly `content`, meaning a file watcher
+ * observing this change is seeing this extension's own write rather than an
+ * external edit.
+ * @param filePath Absolute path to the file that changed.
+ * @param content The file's current content.
  */
-function createFileWatcher(filePath: string): vscode.FileSystemWatcher {
-  const pattern = new vscode.RelativePattern(
-    path.dirname(filePath),
-    path.basename(filePath),
-  );
-  return vscode.workspace.createFileSystemWatcher(pattern);
+export function isManagedFileSelfWrite(
+  filePath: string,
+  content: string,
+): boolean {
+  return selfWriteTracker.isSelfWrite(filePath, content);
 }
 
 /**
@@ -91,7 +84,7 @@ function getUserDirectory(context: vscode.ExtensionContext): string {
  * @param context Extension context.
  * @returns Returns the path to the global storage JSON file.
  */
-function getGlobalStoragePath(context: vscode.ExtensionContext): string {
+export function getGlobalStoragePath(context: vscode.ExtensionContext): string {
   return path.resolve(context.globalStorageUri.fsPath, "../storage.json");
 }
 
@@ -181,7 +174,7 @@ function findByKeyValuePair(
  * @param context Extension context.
  * @returns Returns the name of the current profile.
  */
-async function getCurrentProfileName(
+export async function getCurrentProfileName(
   context: vscode.ExtensionContext,
 ): Promise<string> {
   const storage = await readGlobalStorage(context);
@@ -258,7 +251,7 @@ async function getCurrentProfileName(
  * @param context Extension context.
  * @returns A mapping from profile name to the directory for the profile.
  */
-async function getProfileMap(
+export async function getProfileMap(
   context: vscode.ExtensionContext,
 ): Promise<Record<string, string>> {
   const map: Record<string, string> = {};
@@ -288,7 +281,7 @@ async function getProfileMap(
  * @param context Extension context.
  * @returns Returns details about the current profile.
  */
-async function getCurrentProfileDetails(
+export async function getCurrentProfileDetails(
   context: vscode.ExtensionContext,
 ): Promise<{
   currentProfileName: string;
@@ -493,7 +486,9 @@ async function writeInheritedSettings(
 /**
  * Reads and returns a raw `settings.json` file.
  */
-async function readRawSettingsFile(settingsPath: string): Promise<string> {
+export async function readRawSettingsFile(
+  settingsPath: string,
+): Promise<string> {
   try {
     return await fs.readFile(settingsPath, "utf8");
   } catch (error) {
@@ -684,185 +679,4 @@ export async function removeCurrentProfileInheritedSettings(
       "Inherited settings removed from current profile!",
     );
   }
-}
-
-/**
- * Updates the inherited settings when the profile changes.
- * @param context Extension context.
- */
-export async function updateInheritedSettingsOnProfileChange(
-  context: vscode.ExtensionContext,
-) {
-  const globalStoragePath = getGlobalStoragePath(context);
-  let currentProfile = await getCurrentProfileName(context);
-
-  const watcher = createFileWatcher(globalStoragePath);
-  const onChange = async () => {
-    const newProfileName = await getCurrentProfileName(context);
-    if (newProfileName !== currentProfile) {
-      currentProfile = newProfileName;
-      console.info(
-        "Current profile has changed, updating inherited settings...",
-      );
-      await updateCurrentProfileInheritance(context);
-    }
-  };
-  watcher.onDidChange(onChange);
-  watcher.onDidCreate(onChange);
-  watcher.onDidDelete(onChange);
-
-  context.subscriptions.push(watcher);
-}
-
-/**
- * Watches the current profile's `settings.json` file and re-applies profile
- * inheritance whenever it changes for a reason other than this extension's
- * own writes (e.g. the user editing and saving the file).
- *
- * The active profile (and therefore its `settings.json` path) can change at
- * any time, so this function also watches the global storage file used to
- * track the active profile, and re-subscribes to the new profile's
- * `settings.json` whenever it changes.
- * @param context Extension context.
- */
-export async function registerCurrentProfileSaveWatcher(
-  context: vscode.ExtensionContext,
-): Promise<void> {
-  const scheduleReapply = createDebouncedTrigger(() =>
-    updateCurrentProfileInheritance(context),
-  );
-
-  let settingsWatcher: vscode.FileSystemWatcher | undefined;
-
-  const resubscribe = async () => {
-    settingsWatcher?.dispose();
-    settingsWatcher = undefined;
-
-    let currentProfileDirectory: string;
-    try {
-      ({ currentProfileDirectory } = await getCurrentProfileDetails(context));
-    } catch (error) {
-      console.error(
-        "Failed to resolve the current profile directory for the save watcher:",
-        error,
-      );
-      return;
-    }
-
-    const settingsPath = path.join(currentProfileDirectory, "settings.json");
-    const watcher = createFileWatcher(settingsPath);
-    const onChange = async () => {
-      const latestContent = await readRawSettingsFile(settingsPath);
-      if (selfWriteTracker.isSelfWrite(settingsPath, latestContent)) {
-        return; // Our own write; nothing changed from the user's perspective.
-      }
-      scheduleReapply();
-    };
-    watcher.onDidChange(onChange);
-    watcher.onDidCreate(onChange);
-    settingsWatcher = watcher;
-  };
-
-  await resubscribe();
-
-  // The active profile (and therefore the path watched above) can change at
-  // any time, so watch the global storage file used to track it and
-  // re-subscribe to the new profile's settings.json whenever it changes.
-  const globalStoragePath = getGlobalStoragePath(context);
-  const profileWatcher = createFileWatcher(globalStoragePath);
-  profileWatcher.onDidChange(resubscribe);
-  profileWatcher.onDidCreate(resubscribe);
-  profileWatcher.onDidDelete(resubscribe);
-
-  context.subscriptions.push(profileWatcher, {
-    dispose: () => settingsWatcher?.dispose(),
-  });
-}
-
-/**
- * Watches each parent profile's `settings.json` file (as configured via
- * `inheritProfile.parents`) and re-applies inheritance to the current
- * profile whenever any of them changes.
- *
- * Parent profiles are never written to by this extension, so unlike
- * {@link registerCurrentProfileSaveWatcher} there is no self-write to guard
- * against here.
- *
- * The set of watched files is re-resolved whenever the
- * `inheritProfile.parents` configuration changes, or the active profile
- * changes (since the set of known profiles/directories may also have
- * changed).
- * @param context Extension context.
- */
-export async function registerParentProfileSaveWatcher(
-  context: vscode.ExtensionContext,
-): Promise<void> {
-  const scheduleReapply = createDebouncedTrigger(() =>
-    updateCurrentProfileInheritance(context),
-  );
-
-  let parentWatchers: vscode.FileSystemWatcher[] = [];
-
-  const resubscribe = async () => {
-    for (const watcher of parentWatchers) {
-      watcher.dispose();
-    }
-    parentWatchers = [];
-
-    const config = vscode.workspace.getConfiguration("inheritProfile");
-    const parentProfileNames = config.get<string[]>("parents", []);
-    if (parentProfileNames.length === 0) {
-      return;
-    }
-
-    let profiles: Record<string, string>;
-    try {
-      profiles = await getProfileMap(context);
-    } catch (error) {
-      console.error(
-        "Failed to resolve parent profile directories for the save watcher:",
-        error,
-      );
-      return;
-    }
-
-    const parentSettingsPaths = resolveParentSettingsPaths(
-      parentProfileNames,
-      profiles,
-    );
-    for (const settingsPath of parentSettingsPaths) {
-      const watcher = createFileWatcher(settingsPath);
-      watcher.onDidChange(scheduleReapply);
-      watcher.onDidCreate(scheduleReapply);
-      watcher.onDidDelete(scheduleReapply);
-      parentWatchers.push(watcher);
-    }
-  };
-
-  await resubscribe();
-
-  // Re-subscribe whenever the configured list of parent profiles changes.
-  const configWatcher = vscode.workspace.onDidChangeConfiguration((event) => {
-    if (event.affectsConfiguration("inheritProfile.parents")) {
-      void resubscribe();
-    }
-  });
-
-  // The active profile (and therefore which directory each parent profile
-  // name resolves to) can change at any time, and the set of known profiles
-  // can also change, so watch the global storage file used to track both of
-  // those and re-subscribe whenever it changes.
-  const globalStoragePath = getGlobalStoragePath(context);
-  const profileWatcher = createFileWatcher(globalStoragePath);
-  profileWatcher.onDidChange(resubscribe);
-  profileWatcher.onDidCreate(resubscribe);
-  profileWatcher.onDidDelete(resubscribe);
-
-  context.subscriptions.push(configWatcher, profileWatcher, {
-    dispose: () => {
-      for (const watcher of parentWatchers) {
-        watcher.dispose();
-      }
-    },
-  });
 }
